Update extract.py to create the hst-whitelist.txt
Since this tool creates folders which are not listed in the default `hst-whitelist.txt`, following the instructions resulted in a site with several resources which were blocked from public access.  This creates a new `hst-whitelist.txt` which includes all folders.

I've included the file header, since it may provide important information when users are making future changes to the whitelist.  I have mixed feelings about including it as a multiline string in the python source vs creating a new project file which could be copied and amended.  I went with the former, because I am making this change on Github.com and can only edit one file at a time. 
¯\_(ツ)_/¯